refactor: create db module
