+++ tasks/checker/statusChecker.py
@@ -1,3 +1,7 @@
+"""
+csv 파일 기반 각 기관별 상태 확인
+"""
+
 import asyncio
 import aiohttp
 import async_timeout
